split test_interpolation.py into tests/interpolation/test_interpolation.py and tests/interpolation/test_custom_resolvers.py
